fastapi/requirements.txt
fastapi==0.70.0
python-multipart==0.0.5
aiofiles==0.4.0
numpy==1.22.2
opencv_python==4.5.2.52
scikit_image==0.18.2
loguru==0.5.3
onnxruntime==1.8.0
Pillow==8.3.0